backend/requirements.txt
anyio==3.7.0
click==8.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pydantic==1.10.8
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
typing_extensions==4.6.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
